pydanticai/Level1/requirements.txt
pydantic-ai>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
logfire>=0.19.0
fastui>=0.6.0
uvicorn>=0.27.0
fastapi>=0.109.0
pydantic>=2.0.0
python-multipart>=0.0.6